rename: 404-style 파일 병합
